% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/WhereIsMyPeak.R
\name{initRegions}
\alias{initRegions}
\title{initRegions}
\usage{
initRegions(txdb = NULL, mappingTable = NULL, namingTable = NULL)
}
\arguments{
\item{txdb}{txdb annotation data base}

\item{mappingTable}{mappingTable}

\item{namingTable}{namingTable}
}
\value{
list with preparsed TxDB information
}
\description{
initRegions
}
\author{
Thomas Schwarzl
}

