% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/WhereIsMyPeak.R
\name{shuffle}
\alias{shuffle}
\title{Shuffle}
\usage{
shuffle(annotatedPeaks, txdbParsed)
}
\arguments{
\item{annotatedPeaks}{GRanges object with a elementMetadata column 'region'}

\item{txdbParsed}{parsed TxDB infro}
}
\value{
shuffled peaks
}
\description{
Shuffles
}

