% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/WhereIsMyPeak.R
\name{shuffleRegion}
\alias{shuffleRegion}
\title{Shuffle}
\usage{
shuffleRegion(interval, txdbParsed, background = NULL)
}
\arguments{
\item{interval}{To be shuffled. GRanges object With a elementMetadata column 'region' and 'ID'}

\item{txdbParsed}{parsed TxDB infro}

\item{background}{Background. GRanges object with a elementMetadata column 'region' and 'ID'}
}
\description{
Shuffle
}

