% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/WhereIsMyPeak.R
\name{WhereIsMyPeak}
\alias{WhereIsMyPeak}
\title{WhereIsMyPeak}
\usage{
WhereIsMyPeak(intervals, txdbParsed, annotationOrder = c("Exon", "Intron"),
  exonAnnotationOrder = c("CDS", "5' UTR", "3' UTR"),
  typeOrder = c("non_coding", "known_ncrna", "rRNA", "snoRNA", "snRNA",
  "miRNA", "lincRNA", "Mt_rRNA", "Mt_tRNA", "sense_intronic",
  "transcribed_processed_pseudogene", "transcribed_unitary_pseudogene",
  "transcribed_unprocessed_pseudogene", "translated_processed_pseudogene",
  "translated_unprocessed_pseudogene", "unitary_pseudogene",
  "unprocessed_pseudogene", "sense_overlapping", "3prime_overlapping_ncrna",
  "antisense", "IG_C_gene", "IG_D_gene", "IG_J_gene", "IG_V_gene",
  "IG_C_pseudogene", "IG_J_pseudogene", "IG_V_pseudogene",     
  "polymorphic_pseudogene", "misc_RNA", "TEC", "TR_C_gene", "TR_D_gene",
  "TR_J_gene", "TR_J_pseudogene", "TR_V_gene", "TR_V_pseudogene", "pseudogene",
  "processed_pseudogene", "processed_transcript", "LRG_gene", "protein_coding"))
}
\arguments{
\item{intervals}{peaks as GRanges object}

\item{annotationOrder}{character vector of prioritising genomic locations for annotation}

\item{typeOrder}{character vector prioritising gene types if a tied match}

\item{txdb}{txdb annotation data base}

\item{mappingTable}{mappingTable}

\item{namingTable}{namingTable}
}
\value{
list of results
}
\description{
WhereIsMyPeak - Annotate GenomicRanges for gene location
}
\author{
Thomas Schwarzl
}

